@@ -10,30 +10,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.5",
-    "@angular/common": "^19.0.5",
-    "@angular/compiler": "^19.0.5",
-    "@angular/core": "^19.0.5",
-    "@angular/forms": "^19.0.5",
-    "@angular/platform-browser": "^19.0.5",
-    "@angular/platform-browser-dynamic": "^19.0.5",
-    "@angular/router": "^19.0.5",
+    "@angular/animations": "^20.2.3",
+    "@angular/common": "^20.2.3",
+    "@angular/compiler": "^20.2.3",
+    "@angular/core": "^20.2.3",
+    "@angular/forms": "^20.2.3",
+    "@angular/platform-browser": "^20.2.3",
+    "@angular/platform-browser-dynamic": "^20.2.3",
+    "@angular/router": "^20.2.3",
     "@fontsource/andika": "^4.5.11",
     "@fontsource/jetbrains-mono": "^4.5.11",
-    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^19.0.1",
     "@ngneat/until-destroy": "^10.0.0",
     "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.3.3",
     "leaflet": "^1.9.3",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.15.0"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.6",
-    "@angular/cli": "~19.0.6",
-    "@angular/compiler-cli": "^19.0.5",
-    "@angular/localize": "^19.0.5",
+    "@angular/build": "^20.2.1",
+    "@angular/cli": "~20.2.1",
+    "@angular/compiler-cli": "^20.2.3",
+    "@angular/localize": "^20.2.3",
     "@types/jasmine": "~4.3.0",
     "@types/leaflet": "^1.9.0",
     "jasmine-core": "~4.5.0",
@@ -42,6 +42,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~5.6.3"
+    "typescript": "~5.9.2"
   }
 }
