@@ -1,15 +1,17 @@
-import { Component, inject } from '@angular/core';
-import { ReactiveFormsModule, FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { JsonPipe } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { FormBuilder, FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
+import { ActivatedRoute, Router } from '@angular/router';
+import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 
+@UntilDestroy()
 @Component({
   selector: 'loop-reactive-form-state',
   standalone: true,
   imports: [ReactiveFormsModule, JsonPipe],
   template: `
     <div>
-      <h1>Reactive Form State Example</h1>
-
+      <h2>Reactive Form State</h2>
       <form [formGroup]="form" (ngSubmit)="submit()">
         <div class="d-flex align-items-baseline">
           <div class="input-group mb-3 w-25">
@@ -30,8 +32,6 @@ import { JsonPipe } from '@angular/common';
         <button class="btn btn-primary" type="submit" [disabled]="form.invalid">Submit</button>
       </form>
 
-      <hr />
-
       <h2>Form State</h2>
       <p><strong>Value:</strong></p>
       <pre>{{ form.value | json }}</pre>
@@ -41,13 +41,33 @@ import { JsonPipe } from '@angular/common';
       <p><strong>Dirty:</strong> {{ form.dirty }}</p>
       <p><strong>Touched:</strong> {{ form.touched }}</p>
 
+      <hr>
+
+      <h2>Filter working with URL</h2>
+      <div class="input-group mb-3 w-25">
+        <span class="input-group-text">Filter</span>
+        <input [formControl]="filter" type="text" class="form-control" placeholder="filter" aria-label="filter">
+      </div>
+
     </div>
   `,
 })
 export class FormStateComponent {
+  #activeRoute = inject(ActivatedRoute);
+  #router = inject(Router);
+
+  filter = new FormControl(this.#activeRoute.snapshot.params['filter']);
+
+  constructor() {
+    this.filter.valueChanges.pipe(untilDestroyed(this)).subscribe((value) => {
+      this.#router.navigate([{ filter: value }], { queryParamsHandling: 'merge', relativeTo: this.#activeRoute });
+    });
+    //TODO: listen to route changes and filter playgrounds below
+  }
+
   form = inject(FormBuilder).group({
-    name: ['', [Validators.required]],
-    email: ['', [Validators.required, Validators.email]]
+    name: [this.#activeRoute.snapshot.params['name'], [Validators.required]],
+    email: [this.#activeRoute.snapshot.params['email'], [Validators.required, Validators.email]]
   })
 
   submit() {
