@@ -1,6 +1,6 @@
 import { inject, Injectable } from '@angular/core';
-import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { BehaviorSubject, catchError, of, ReplaySubject, shareReplay, switchMap, tap } from 'rxjs';
+import { UntilDestroy } from '@ngneat/until-destroy';
+import { BehaviorSubject, catchError, Observable, of, ReplaySubject, shareReplay, switchMap, tap } from 'rxjs';
 import { Playground } from 'src/app/shared';
 import { RxjsPlaygroundService } from './rxjs-playground.service';
 
@@ -9,39 +9,25 @@ import { RxjsPlaygroundService } from './rxjs-playground.service';
 export class RxjsPlaygroundStore {
 
   #service = inject(RxjsPlaygroundService);
-  #reload = new BehaviorSubject<void>(undefined);
+  #refresh = new BehaviorSubject<void>(undefined);
+  readonly updateError = new BehaviorSubject<Playground | undefined>(undefined);
+  readonly loading = new BehaviorSubject<boolean>(true);
+  readonly error = new ReplaySubject<any>();
 
-  #update = new BehaviorSubject<Playground | undefined>(undefined);
-  #updateSubscription = this.#update.pipe(
-    switchMap(playground => !playground ? of(null) :  this.#service.update(playground.id, playground).pipe(
-        tap(() => this.#reload.next()),
-        catchError(error => {
-          this.updateError.next(error);
-          return of(null);
-        })
-      )
-    ),
-    untilDestroyed(this),
-  ).subscribe();
-
-  readonly updateError = new ReplaySubject<any>();
-
-  readonly playgroundsLoading = new BehaviorSubject<boolean>(true);
-  readonly playgroundsError = new ReplaySubject<any>();
-  readonly playgrounds = this.#reload.pipe(
-    tap(() => this.playgroundsLoading.next(true)),
+  readonly playgrounds = this.#refresh.pipe(
+    tap(() => this.loading.next(true)),
     switchMap(() => this.#service.list()),
-    catchError(error => {
-      this.playgroundsError.next(error);
-      return of([]);
-    }),
-    tap(() => this.playgroundsLoading.next(false)),
-    shareReplay(1)
+    tap(() => this.error.next(undefined)),
+    catchError(error => { this.error.next(error); return of([]) }),
+    tap(() => this.loading.next(false)),
+    shareReplay(1),
   );
 
-
-  update(playground: Playground) {
-    this.#update.next(playground);
+  update(playground: Playground): Observable<void> {
+    return this.#service.update(playground).pipe(
+      catchError(error => { this.updateError.next(error); throw error; }),
+      tap(() => { this.#refresh.next(); this.updateError.next(undefined); }),
+    );
   }
 }
 
