@@ -1,6 +1,6 @@
 import { inject, signal } from "@angular/core";
 import { rxResource } from "@angular/core/rxjs-interop";
-import { of } from "rxjs";
+import { of, tap } from "rxjs";
 import { Playground } from "../model";
 import { PlaygroundService } from "../service";
 
@@ -12,23 +12,25 @@ export class PlaygroundStore {
   });
   #updateResource = rxResource({
     request: () => this.#updatePlayground(),
-    loader: ({ request }) => request ? this.#service.update(request.id, request) : of(undefined)
+    loader: ({ request }) => request ? this.#service.update(request.id, request).pipe(
+      tap(() => this.#playgroundsResource.reload())
+    ) : of(undefined)
   });
-
   #deleteResource = rxResource({
     request: () => this.#deletePlayground(),
-    loader: ({ request }) => request ? this.#service.delete(request) : of(undefined)
+    loader: ({ request }) => request ? this.#service.delete(request).pipe(
+      tap(() => this.#playgroundsResource.reload())
+    ) : of(undefined)
   });
 
-  playgrounds = this.#playgroundsResource.value;
-  playgroundsError = this.#playgroundsResource.error;
-  loading = this.#playgroundsResource.isLoading;
+  readonly playgrounds = this.#playgroundsResource.value.asReadonly();
+  readonly playgroundsError = this.#playgroundsResource.error;
+  readonly loading = this.#playgroundsResource.isLoading;
+  readonly updateError = this.#updateResource.error;
+  readonly deleteError = this.#deleteResource.error;
 
   #updatePlayground = signal<Playground | null>(null);
   #deletePlayground = signal<string | null>(null);
-
-  updateError = this.#updateResource.error;
-  deleteError = this.#deleteResource.error;
 
   update(playground: Playground): void {
     this.#updatePlayground.set(playground);
