@@ -8,13 +8,12 @@ import { Component } from '@angular/core';
 })
 export class NgforComponent {
 
-  items: string[] = ['foo', 'bar'];
+  items: string[] = ['1. foo', '2. bar'];
 
   myTrack = (index: number, text: string) => text;
 
   addInput(input: string) {
-    this.items.push(input);
-    this.items.sort();
+    this.items.push(`${this.items.length + 1}. ${input}`);
   }
 
 }
