@@ -1,26 +1,23 @@
-import { NgFor } from '@angular/common';
+
 import { Component } from '@angular/core';
-import { AbstractControl, FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { Person } from '../person';
 
-export type TypedForm<T> = { [P in keyof T]?: AbstractControl<T[P] | null> };
-
 @Component({
-    selector: 'app-typed-form',
-    templateUrl: './typed-form.component.html',
-    imports: [
-        ReactiveFormsModule,
-        NgFor,
-    ]
+  selector: 'app-typed-form',
+  templateUrl: './typed-form.component.html',
+  imports: [
+    ReactiveFormsModule
+  ]
 })
 export class TypedFormComponent {
 
   colors = ["Red", "Green", "Blue"];
   model = new Person('Flemming', 'Bregnvig', "Blue", 182);
 
-  fg = this.fb.group<TypedForm<Person>>({
+  fg = this.fb.group({
     firstName: this.fb.control<string>(this.model.firstName, [Validators.required]),
-    lastName: this.fb.control(this.model.firstName, [Validators.required]),
+    lastName: this.fb.control(this.model.lastName, [Validators.required]),
     favoriteColor: this.fb.control(this.model.favoriteColor),
     height: this.fb.control(this.model.height, [Validators.min(100), Validators.max(200)])
   });
