@@ -1,13 +1,13 @@
 import { Component, QueryList, ViewChildren } from '@angular/core';
 
 import { StopwatchComponent } from '../stopwatch';
-import { NgFor } from '@angular/common';
+
 import { StopwatchComponent as StopwatchComponent_1 } from '../stopwatch/stopwatch.component';
 
 @Component({
     selector: 'app-view-children',
     templateUrl: './view-children.component.html',
-    imports: [StopwatchComponent_1, NgFor]
+    imports: [StopwatchComponent_1]
 })
 export class ViewChildrenComponent {
   @ViewChildren(StopwatchComponent) stopwatches?: QueryList<StopwatchComponent>;
