@@ -1,14 +1,14 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { Menu } from '../menu.model';
 import { RouterLinkActive, RouterLink } from '@angular/router';
-import { NgFor } from '@angular/common';
+
 
 @Component({
     selector: 'examples-menu',
     templateUrl: './menu.component.html',
     styleUrls: ['./menu.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [NgFor, RouterLinkActive, RouterLink]
+    imports: [RouterLinkActive, RouterLink]
 })
 export class MenuComponent {
 
