@@ -1,8 +1,8 @@
-import { AsyncPipe } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, computed, Signal } from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Center, LeafletModule, Marker } from '@loopme/leaflet';
-import { Observable, combineLatest, map, shareReplay, switchMap } from 'rxjs';
+import { map, switchMap } from 'rxjs';
 import { FooterComponent } from '../footer/footer.component';
 import { LocationService, Playground } from '../shared';
 import { PlaygroundService } from '../shared/playground.service';
@@ -12,26 +12,25 @@ import { SidebarComponent } from '../sidebar/sidebar.component';
   selector: 'app-map',
   template: `
     <main class="vh-100 d-flex justify-content-center align-items-center">
-      <leaflet [center]="center" [markers]="markers$ | async"></leaflet>
+      <leaflet [center]="center()" [markers]="markers()"></leaflet>
     </main>
-    <app-sidebar [playgrounds]="appPlaygrounds" [selectedPlayground]="playground" (selected)="playgroundSelected($event)"></app-sidebar>
-    @if(playground) {
-      <app-footer [playground]="playground"></app-footer>
+    <app-sidebar [playgrounds]="appPlaygrounds()" [selectedPlayground]="playground()" (selected)="playgroundSelected($event)"/>
+    @if(playground(); as p) {
+      <app-footer [playground]="p"/>
     }
   `,
   imports: [
     LeafletModule,
     SidebarComponent,
     FooterComponent,
-    AsyncPipe,
   ],
 })
 export class MapComponent {
 
-  appPlaygrounds?: Playground[];
-  playground?: Playground;
-  center: Center = { lat: 56.360029, lng: 10.746635 };
-  markers$: Observable<Marker[]>;
+  appPlaygrounds: Signal<Playground[]>;
+  playground: Signal<Playground | undefined>;
+  center: Signal<Center>;
+  markers: Signal<Marker[]>;
 
   constructor(
     service: PlaygroundService,
@@ -39,19 +38,24 @@ export class MapComponent {
     route: ActivatedRoute,
     private router: Router) {
 
-    service.getPlaygrounds().subscribe(playgrounds => this.appPlaygrounds = playgrounds);
-    const playground$ = route.params.pipe(
-      map(params => params['id']),
-      switchMap(id => service.find(id)),
-      shareReplay(1),
-    );
-    this.markers$ = combineLatest([
-      locationService.current,
-      playground$.pipe(map(p => p?.position))
-    ]);
-    playground$.subscribe(playground => {
-      this.playground = playground;
-      playground && (this.center = { ...playground.position, zoom: 14 });
+    this.appPlaygrounds = toSignal(service.getPlaygrounds(), { initialValue: [] });
+
+    this.playground = toSignal<Playground | undefined>(
+      route.params.pipe(
+        map(params => params['id']),
+        switchMap(id => service.find(id)),
+      ));
+    this.markers = computed(() => {
+      const current = locationService.current();
+      const playground = this.playground();
+      return [
+        current,
+        playground?.position
+      ];
+    });
+    this.center = computed(() => {
+      const playground = this.playground();
+      return playground?.position ? { ...playground.position, zoom: 12 } : { lat: 56.360029, lng: 10.746635, zoom: 7 };
     });
   }
 
